requirements.txt
opencv-2
mediapipe